+++ src/main/kotlin/com/sesac/climb_mates/account/LoginController.kt
@@ -1,0 +1,12 @@
+package com.sesac.climb_mates.account
+
+import org.springframework.stereotype.Controller
+import org.springframework.web.bind.annotation.GetMapping
+
+@Controller
+class LoginController {
+    @GetMapping("/login")
+    fun loginPage(): String {
+        return "account/login"
+    }
+}
