+++ src/main/kotlin/com/sesac/climb_mates/config/GlobalControllerAdvice.kt
@@ -1,4 +1,4 @@
-package com.sesac.climb_mates.controller
+package com.sesac.climb_mates.config
 
 import jakarta.servlet.http.HttpSession
 import org.springframework.web.bind.annotation.ControllerAdvice
