--- src/main/kotlin/com/sesac/climb_mates/controller/IndexController.kt
@@ -1,22 +1,0 @@
-package com.sesac.climb_mates.controller
-
-import org.springframework.stereotype.Controller
-import org.springframework.web.bind.annotation.GetMapping
-
-@Controller
-class IndexController(
-
-) {
-    @GetMapping("/")
-    fun indexPage():String{
-        return "main/main"
-    }
-    @GetMapping("/login")
-    fun loginPage():String{
-        return "account/login"
-    }
-    @GetMapping("/test")
-    fun testPage(): String {
-        return "test";
-    }
-}
