+++ src/main/kotlin/com/sesac/climb_mates/error/CustomErrorController.kt
@@ -1,4 +1,4 @@
-package com.sesac.climb_mates.controller
+package com.sesac.climb_mates.error
 
 import jakarta.servlet.RequestDispatcher
 import jakarta.servlet.http.HttpServletRequest
