+++ src/main/kotlin/com/sesac/climb_mates/grouping/data/applicant/GroupingApplicant.kt
@@ -1,25 +1,23 @@
-package com.sesac.climb_mates.grouping.data
+package com.sesac.climb_mates.grouping.data.applicant
 
 import com.sesac.climb_mates.account.data.Account
+import com.sesac.climb_mates.grouping.data.Grouping
 import jakarta.persistence.*
 import java.time.LocalDateTime
 
 @Entity
-@Table(name="grouping_review")
-data class GroupingReview(
+@Table(name="grouping_applicant")
+data class GroupingApplicant(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id:Long? = null,
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "grouping_id", nullable = false)
     var grouping: Grouping,
+
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "account_id", nullable = false)
     var account: Account,
-
-    @Column(name="content", nullable = false)
-    var content:String,
 
     @Column(name="created_date", nullable = false)
     val createdDate: LocalDateTime = LocalDateTime.now()
